Legacy cleanup, part 2 (#3659)
* Legacy cleanup, part 2

There are still things that can be removed, those will be likely done in part three:
`The Return of the Cleanup` :)


* Fix windows build

Somehow there were some unfixed errors from enabled warnings?
I'm unsure if I've pulled repo in unstable state, or those were somehow
missed.

+++ Gems/LyShine/Code/Editor/EditorWindow.cpp
@@ -2112,8 +2112,8 @@ int EditorWindow::GetCanvasMaxHierarchyDepth(const LyShine::EntityArray& rootChi
         return depth;
     }
 
-    int numChildrenCurLevel = rootChildElements.size();
-    int numChildrenNextLevel = 0;
+    size_t numChildrenCurLevel = rootChildElements.size();
+    size_t numChildrenNextLevel = 0;
     std::list<AZ::Entity*> elementList(rootChildElements.begin(), rootChildElements.end());
     while (!elementList.empty())
     {
